+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* groovydocJar use javadoc classifier if no java sources available (because maven central requires javadoc for publication)
+* generate pom.properties inside jar MEAT-INF (like maven)
+
 ### 1.0.0 (2015-11-23)
 
 * Initial release